Outline fixed and started working on tetrominos again

+++ src/BoardOutline.cpp
@@ -1,6 +1,6 @@
 #include "classes.h"
 
-Board::Outline::Outline(int x, int y){
+Board::Outline::Outline(float x, float y){
 	posx = x;
 	posy = y;
 	body = {10,10,10,10};
@@ -8,7 +8,7 @@ Board::Outline::Outline(int x, int y){
 }
 //setter to the outline
 void Board::Outline::setOutline(Color color,float ccountw,float ccounth){
-	body = {(float)posx,(float)posy,ccountw,ccounth};
+	body = {posx,posy,ccountw,ccounth};
 	c = color;
 }
 void Board::Outline::DrawOutline(float linesize){
